fix!: bug 4157, use getAccounts on HD Keyring when calling addNewAccount (#4158)
## Explanation

This PR addresses a bug in the keyring controller where invoking
addNewAccount inadvertently retrieves addresses from all keyrings,
rather than exclusively from the HD keyring. This behavior leads to
inconsistencies and errors when accounts from different keyring types
are present. The primary symptom of this bug is the retrieval of an
incorrect number of accounts, with some operations potentially resulting
in an undefined address. This issue is notably impactful as it can cause
the selectedAddress to be undefined in clients, leading to unexpected
behavior and degraded user experience.

## References

* Fixes #4157

## Changelog

### `@metamask/keyring-controller`

- **BREAKING**: Use getAccounts on HD keyring when calling
`addNewAccount` instead of `this.getAccounts` which retrieves every
account.

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [ ] I've highlighted breaking changes using the "BREAKING" category
above as appropriate