Fix `getProviderAndBlockTracker` client selection logic (#4187)
## Explanation

<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?
* Are there any changes whose purpose might not obvious to those
unfamiliar with the domain?
* If your primary goal was to update one package but you found you had
to update another one along the way, why did you do so?
* If you had to upgrade a dependency, why did you do so?
-->

This change fixes a bug where if `getProviderAndBlockTracker` is called
with a domain that 1) does not have permissions, 2) does not have a
cached proxy and 3) the `useRequestQueue` preference is on, it's newly
created networkProxy will not be pointed correctly. Under these
conditions the network proxy for this domain should point to the
NetworkController's own proxied selectedNetworkClient.

I've added tests to ensure that all possible permutations are pointing
the proxy correctly:
e.g. 
- useRequestQueue preference on + domain does not have permissions 
- useRequestQueue preference off + domain does have permissions
- useRequestQueue preference on + domain does have permissions
- etc

## Changelog

### `@metamask/selected-network-controller`

- **FIXED**: When `getProviderBlockTracker` is called with domain for
which there is no cached `networkProxy` in the `domainProxyMap`, if the
`useRequestQueue` preference is off and the domain does not have
permissions the newly created network proxy for this domain should be
pointed at the `NetworkController`'s own proxied
`selectedNetworkClient`.

## Checklist

- [ ] I've updated the test suite for new or updated code as appropriate
- [ ] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [ ] I've highlighted breaking changes using the "BREAKING" category
above as appropriate