docs: Remove various TypeScript@4.4 TODOs
Prior to TypeScript@4.4, I added various TODOs about marking certain
object interfaces properties as optional instead of `type | null`. Now,
years later and a long time after TS4.4 shipped, I have concluded that
it's not worth the trouble. The same goes for a TODO exhorting us to
enable `noUncheckedIndexedAcces` in the PermissionController, which
appears to be something we have to do in a monorepo-wide fashion, if the
test failures it produced are any indication.