Add nonRPCGasFeeApisDisabled property to the gas fee controller (#4094)
## Explanation

This PR allows the user of this controller to control whether it makes
any network requests to extermal APIs other than the rpc provider. To do
this it adds a `nonRPCGasFeeApisDisabled` property to the controller
state. This property defaults to `false`. Methods to set that property
to true and false are added

If that property is `true`, then `determineGasFeeCalculations` will not
make requests to the
`legacyAPIEndpoint`/`fetchLegacyGasPriceEstimatesUrl` nor the
`EIP1559APIEndpoint`/`fetchGasEstimatesUrl`. It will, in this case, only
get gas data via `eth_feeHistory` or `eth_gasPrice` requests to the rpc
provider.

The goal is to allow the user of the controller to prevent network
requests to additional parties, which may be desired to allow end users
greater control over privacy.

## References

This is part of resolving
https://github.com/MetaMask/MetaMask-planning/issues/2076

## Changelog

### `@metamask/gas-fee-controller`

- ADDED: Adds nonRPCGasFeeApisDisabled, property to the controller,
allowing the user to prevent requests for gas fees to the external
'APIEndpoints'

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate