fix: Restore subjectTypes test
One of the `subjectTypes` test cases was commented because Jest
encountered a type error when attempting to run it. This test case is
now restored by re-implementing the messenger action calling spy, and
relaxing the expectations thereon. (The exact number and order of
action calls differ between incremental and non-incremental permission
requests.)