Add way to view pkg changes since latest release
When creating a new release, it can be helpful to be able to view the
changes that have been made to a package since its most recent release.
This commit adds a package script, `since-latest-release`, which can be
used to accomplish this.

This script is simply a shortcut for a commit range, so you can plug any
Git command that takes a commit range into it. For instance, to view the
list of latest commits for `@metamask/assets-controllers`, you can use
`log`:

    yarn workspace @metamask/assets-controllers since-latest-release log

Or to view all changes, you can use `diff`:

    yarn workspace @metamask/assets-controllers since-latest-release diff

Or you can leave an argument off to just get the names of commits:

    yarn workspace @metamask/assets-controllers since-latest-release