Refactor TokensController tests
We want to change TokensController so that it relies on
`selectedNetworkClientId` instead of `providerConfig` to know which
chain is selected, and this will require updating the tests to match.
However, the TokensController tests are written in such a way that makes
updating them difficult.

This commit rewrites each test to use the `withController` pattern that
we've adopted for other controller tests. It also removes usage of
`beforeEach` so that each test is isolated.

Note that this commit isn't a complete refactor of the tests; there is
more work we could do to make them easier to maintain, such as removing
a reliance on ApprovalController and using `describe` to group tests by
method. This step can be done later.

In addition, this commit also adds some NetworkController-related and
BaseController-related test helpers which the TokensController tests
use. These helpers aren't defined in the public part of their respective
packages, but in the private area. The TokensController tests can import
them directly because they are private themselves, and therefore don't
need to adhere to the same importing rules that the (public)
implementation files do.

Finally, this commit also exposes `BuiltInNetworkClientId` and
`CustomNetworkClientId`. These are used in the aforementioned test
helpers but are also useful on their own.