feat: add patch code from PR 23903 to keyringController. (#4256)
This PR apply patch code from this
[PR](https://github.com/MetaMask/metamask-extension/pull/23903) to
keyringController to fix the issue found in this
[issue](https://github.com/MetaMask/metamask-extension/issues/23804)


## References
[PR 23903](https://github.com/MetaMask/metamask-extension/pull/23903)
Fix [Issue
23804](https://github.com/MetaMask/metamask-extension/issues/23804)

<!--
Are there any issues that this pull request is tied to? Are there other
links that reviewers should consult to understand these changes better?

For example:

* Fixes #12345
* Related to #67890
-->

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/keyring-controlller`

- **Fixed**: Fix creation of new QR keyrings
- Previously the QR keyring would get initialized with invalid state
when created for the first time


## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate