Bump Prettier to 3
Prettier 3 has the ability to read from both `.gitignore` and
`.prettierignore`. This allows us to finetune what we exclude from
Prettier vs. Git in a consistent manner (using configuration files
instead of passing arguments to `prettier`). We will make use of this in
a future commit.

However as v3 makes some backward-incompatible changes to the formatting
rules, this commit applies fixes to all files so that they pass lint
again.