fix: Fix `profile-sync-controller` version
The version of the `@metamask/profile-sync-controller` package has been
updated to `0.0.0`, which is what our publishing automation expects of
all unpublished public packages.

The current version is `0.1.0`, which caused problems for our most
recent release, requiring a rollback. It broke publishing because our
publishing scripts tried to find version `0.1.0` in the package
changelog but were unable to.

This should have been caught by our linter, but it wasn't due to a
deficiency in the `@metamask/auto-changelog` validation command (it
doesn't validate that the current release exists). This problem has
been documented here: https://github.com/MetaMask/auto-changelog/issues/192