feat(keyring-controller): allow non-evm addresses for non-evm methods/functions (#4282)
## Explanation

This allows support of non-EVM addresses for most keyring methods.

Right now, our addresses are just plain `string`, meaning we have not
easy way to detect if an address is an ethereum one or not. So we use
runtime check to verify this, and based on this we adapt some logic
(mostly the address normalization right now).

Also, some `Hex` were being dropped in favor of `string` type to make
addresses a bit more "generic".

## References

<!--
Are there any issues that this pull request is tied to? Are there other
links that reviewers should consult to understand these changes better?

For example:

* Fixes #12345
* Related to #67890
-->

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->


### `@metamask/keyring-controller`

- **ADDED**: Basic support for non-EVM addresses

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate