feat: add swap+send type (#4298)
## Explanation

### What is the current state of things and why does it need to change?

With https://github.com/MetaMask/metamask-extension/pull/23703, we will
be implementing a new type of transaction, swap+send – a swap targeting
the recipient. To properly track this transaction, we will need a new
transaction type with new properties.

### What is the solution your changes offer and how does it work?

We add a new `swapAndSend` transaction type with properties that track
the source+destination tokens, as well as the recipient.

### Are there any changes whose purpose might not obvious to those
unfamiliar with the domain?

We include certain swap properties that we may not use initially (e.g.,
`swapTokenValue`, `swapMetaData`) in case we add more parity with the
swaps functionality over time.

### If your primary goal was to update one package but you found you had
to update another one along the way, why did you do so?

The `TransactionType` enum was abstracted to this library rather than
the Extension.

### If you had to upgrade a dependency, why did you do so?

N/A

## References

[MB-843](https://consensyssoftware.atlassian.net/browse/METABRIDGE-843)

https://github.com/MetaMask/metamask-extension/pull/23703
(https://github.com/MetaMask/metamask-extension/pull/24429 latest in
merge train targeting that branch)

<!--
Are there any issues that this pull request is tied to? Are there other
links that reviewers should consult to understand these changes better?


For example:

* -->

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/transaction-controller`

- **ADDED**: `swapAndSend` type and respective 
- **ADDED**: `updateSwapAndSendTransaction()`
- **ADDED**: `TransactionMetaBase` properties: `destinationTokenAmount`,
`sourceTokenAddress`, `sourceTokenAmount`, `sourceTokenDecimals`,
`swapAndSendRecipient`

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate