Feat/preferences controller smart tx opt in (#3815)
## Explanation

<!--
Thanks for your contribution! Take a moment to answer these questions so
that reviewers have the information they need to properly understand
your changes:

* What is the current state of things and why does it need to change?
* What is the solution your changes offer and how does it work?
* Are there any changes whose purpose might not obvious to those
unfamiliar with the domain?
* If your primary goal was to update one package but you found you had
to update another one along the way, why did you do so?
* If you had to upgrade a dependency, why did you do so?
-->

This PR introduces a new state property `smartTransactionsOptInStatus`
to the `PreferencesController`. This property allows the user to opt in
or opt out of smart transactions.

This is to match changes introduced by the patch in Mobile
https://github.com/MetaMask/metamask-mobile/pull/9448

**Changes:**

1. Added `smartTransactionsOptInStatus` to the `PreferencesController`
state.
2. Added a new method `smartTransactionsOptInStatus` to update the
`smartTransactionsOptInStatus` state.
3. Updated the `PreferencesController` tests to include tests for
`smartTransactionsOptInStatus` and to check the default state of
`smartTransactionsOptInStatus`.

**Test Plan:**

1. Run the updated tests for `PreferencesController` using the command
`yarn workspace @metamask/preferences-controller run test`.
2. Ensure all tests pass successfully.

## References

Related to https://github.com/MetaMask/metamask-mobile/pull/8337

<!--
Are there any issues that this pull request is tied to? Are there other
links that reviewers should consult to understand these changes better?

For example:

* Fixes #12345
* Related to #67890
-->

## Changelog

<!--
If you're making any consumer-facing changes, list those changes here as
if you were updating a changelog, using the template below as a guide.

(CATEGORY is one of BREAKING, ADDED, CHANGED, DEPRECATED, REMOVED, or
FIXED. For security-related issues, follow the Security Advisory
process.)

Please take care to name the exact pieces of the API you've added or
changed (e.g. types, interfaces, functions, or methods).

If there are any breaking changes, make sure to offer a solution for
consumers to follow once they upgrade to the changes.

Finally, if you're only making changes to development scripts or tests,
you may replace the template below with "None".
-->

### `@metamask/preferences-controller`

- **<ADDED>** Added preference to enable smart transactions

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [ ] I've highlighted breaking changes using the "BREAKING" category
above as appropriate