TokenRatesController: providerConfig -> selectedNetworkClientId (#4317)
The `providerConfig` state property is being removed from
NetworkController. Currently this property is used in
TokenRatesController to get the currently selected chain and ticker, but
`selectedNetworkClientId` can be used instead to get the currently
selected network client, and then the appropriate information can be
read from that object. This commit makes that transition so that we can
fully drop `providerConfig`.