Modify build process to disable code splitting
The code splitting feature in `tsup` breaks up source code into logical
modules so that if a module is not being used it can be removed by a
build tool that supports tree shaking.

However, the modules that `tsup` creates do not necessary match the
files in the original source code, and `tsup` gives them a random name.
This makes it frustrating for developers on a client team who are
debugging a `core` package by inspecting the compiled version, because
it is impossible to tell what a chunk file contains without first
opening it up.

This commit thus configures `tsup` to disable the code splitting
feature.