feat: Implement market data by Chain ID (#4206)
## Explanation

This pull request introduces a significant enhancement to our
application's functionality by integrating a new feature that presents
market data more comprehensively. This includes updates such as daily
percentage changes and the fluctuation in amounts for a variety of
assets. To accommodate this, we've expanded our state structure to
encapsulate a broader array of data points. Below is a snippet
illustrating the enriched data now available:
```
{
    "0x3845badade8e6dff049820680d1f14bd3903a5d0": {
        "id": "the-sandbox",
        "price": 0.00012453786505819405,
        "marketCap": 282288.57927502826,
        ...
        "pricePercentChange1y": -9.725813860523756
    },
    "0xdac17f958d2ee523a2206206994597c13d831ec7": {
        "id": "tether",
        "price": 0.00026894681330590396,
        "marketCap": 30060069.165317584,
        ...
        "pricePercentChange1y": -0.05851492115750438
    }
}
```

As a result, the TokenRatesController state has been meticulously
updated to include this marketData, enriching our application with a
more detailed and dynamic representation of asset performance. This
enhancement not only broadens the scope of information we provide but
also elevates the user experience by offering a granular view of market
trends and asset valuations.
## References

## Changelog

We've introduced a new marketData property to the application's state,
providing a comprehensive overview of market-related information for
various assets. This enhancement brings a wealth of data directly into
our application.
This update significantly enriches the data available within our
application, offering users detailed insights into asset performance and
market trends. The marketData property is meticulously structured to
ensure easy access to a wide array of information, enhancing the overall
user experience by providing a granular and dynamic view of the market.


### `@metamask/package-assets-controllers`

- **ADDED**: Added `marketData`  to TokenRatesState 
- **REMOVED**: Removed `contractExchangeRates` from TokenRatesState
- **REMOVED**: Removed `contractExchangeRatesByChainId` from
TokenRatesState

## Checklist

- [x] I've updated the test suite for new or updated code as appropriate
- [x] I've updated documentation (JSDoc, Markdown, etc.) for new or
updated code as appropriate
- [x] I've highlighted breaking changes using the "BREAKING" category
above as appropriate